Rename model_db > db
